pyproject.toml
[tool.poetry]
name = "sdXray"
version = "0.1.0"
description = "Survival Analysis on X-Ray image dataset"
authors = ["Nagi0 <brunolefonseca13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
protobuf = "<=3.20"
scikit-learn = "^1.5.2"
opencv-python = "^4.10.0.84"
tqdm = "^4.67.1"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
scikit-image = "^0.24.0"
numpy = "<2.0"
tensorflow-io-gcs-filesystem = "0.27.0"
tensorflow = "<2.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
